feat(firebase): enhance Firebase initialization process
- Updated FirebaseInitializer to initialize Firestore, Remote Config, and App Check.
- Added support for emulator settings for Firestore and Realtime Database.
- Improved logging for Firebase Database in debug mode.
- Removed unnecessary metadata from AndroidManifest.xml.
