+++ core/data/src/main/AndroidManifest.xml
@@ -18,12 +18,6 @@
             <meta-data
                 android:name="com.keelim.core.data.initialize.FirebaseInitializer"
                 android:value="androidx.startup" />
-            <meta-data
-                android:name="com.keelim.core.data.initialize.FirebaseRemoteConfigInitializer"
-                android:value="androidx.startup" />
-            <meta-data
-                android:name="com.keelim.core.data.initialize.FirebaseDatabaseInitializer"
-                android:value="androidx.startup" />
         </provider>
     </application>
 </manifest>
