--- core/data/src/main/java/com/keelim/core/data/initialize/FirebaseRemoteConfigInitializer.kt
@@ -1,24 +1,0 @@
-package com.keelim.core.data.initialize
-
-import android.content.Context
-import androidx.startup.Initializer
-import com.google.firebase.Firebase
-import com.google.firebase.remoteconfig.FirebaseRemoteConfig
-import com.google.firebase.remoteconfig.remoteConfig
-import com.google.firebase.remoteconfig.remoteConfigSettings
-
-class FirebaseRemoteConfigInitializer : Initializer<FirebaseRemoteConfig> {
-    override fun create(context: Context): FirebaseRemoteConfig {
-        return Firebase.remoteConfig.apply {
-            val configSettings = remoteConfigSettings {
-                minimumFetchIntervalInSeconds = 600
-            }
-            setConfigSettingsAsync(configSettings)
-            fetchAndActivate()
-        }
-    }
-
-    override fun dependencies(): List<Class<out Initializer<*>?>?> = listOf(
-        FirebaseInitializer::class.java,
-    )
-}
