--- core/data/src/main/java/com/keelim/core/data/initialize/FirestoreInitializer.kt
@@ -1,40 +1,0 @@
-package com.keelim.core.data.initialize
-
-import android.content.Context
-import androidx.startup.Initializer
-import com.google.firebase.Firebase
-import com.google.firebase.database.database
-import com.google.firebase.firestore.firestore
-import com.google.firebase.firestore.firestoreSettings
-import com.google.firebase.firestore.memoryCacheSettings
-import com.google.firebase.firestore.persistentCacheSettings
-
-internal const val isEmulatorEnable = false
-
-class FirestoreInitializer : Initializer<Unit> {
-
-    override fun create(context: Context) {
-        val firestore = Firebase.firestore
-        if (isEmulatorEnable) {
-            // 10.0.2.2 is the special IP address to connect to the 'localhost' of
-            // the host computer from an Android emulator.
-            firestore.useEmulator("10.0.2.2", 8080)
-
-            // 10.0.2.2 is the special IP address to connect to the 'localhost' of
-            // the host computer from an Android emulator.
-            val database = Firebase.database
-            database.useEmulator("10.0.2.2", 9000)
-        }
-
-        firestore.firestoreSettings = firestoreSettings {
-            // Use memory cache
-            setLocalCacheSettings(memoryCacheSettings {})
-            // Use persistent disk cache (default)
-            setLocalCacheSettings(persistentCacheSettings {})
-        }
-    }
-
-    override fun dependencies(): List<Class<out Initializer<*>>> = listOf(
-        FirebaseInitializer::class.java,
-    )
-}
